requirements.txt
onetimepass
python-gnupg
docopt

